+++ src/app/about/page.tsx
@@ -1,7 +1,6 @@
 "use client"
-import React, { useEffect, useRef, useState } from "react"
-import emailjs from '@emailjs/browser';
-import { Box, Button, Divider, Link, Stack, TextField, Typography } from "@mui/material"
+import React from "react"
+import { Box, Divider, Link, Stack, Typography } from "@mui/material"
 import Grid from "@mui/material/Grid"
 import Image from "next/image"
 import encodeEncyclopedia from "../../../public/assets/about/images/encodeencyclopedia.png"
@@ -9,83 +8,9 @@ import classifications from "../../../public/assets/about/images/classifications
 import biosamples from "../../../public/assets/about/images/biosamples.png"
 import { CA_CTCF, CA_H3K4me3, CA_TF, CA_only, PLS, TF_only, dELS } from "../../common/lib/colors"
 import { CreateLink } from "../../common/utility";
+import ContactForm from "./contactForm";
 
 export default function About() {
-  const [contactName, setContactName] = useState('')
-  const [contactEmail, setContactEmail] = useState('')
-  const [contactMessage, setContactMessage] = useState('')
-  const [error, setError] = useState({ name: false, email: false, message: false })
-  const [success, setSuccess] = useState(false)
-
-  const form = useRef<HTMLFormElement>(null);
-
-  function isValidEmail(email: string): boolean {
-    //hopefully this is right, got it from ChatGPT
-    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-    return emailRegex.test(email);
-  }
-
-  // Move validation logic into useEffect to avoid state updates during render
-  useEffect(() => {
-    const newError = { ...error };
-    let hasChanges = false;
-
-    if (error.email && isValidEmail(contactEmail)) {
-      newError.email = false;
-      hasChanges = true;
-    }
-    if (error.name && contactName) {
-      newError.name = false;
-      hasChanges = true;
-    }
-    if (error.message && contactMessage) {
-      newError.message = false;
-      hasChanges = true;
-    }
-
-    if (hasChanges) {
-      setError(newError);
-    }
-  }, [contactEmail, contactName, contactMessage, error.email, error.name, error.message, error]);
-
-  const handleSubmit = async () => {
-    //Check fields to see if valid
-    const newErrorState = { name: false, email: false, message: false };
-    if (!contactName) newErrorState.name = true;
-    if (!isValidEmail(contactEmail)) newErrorState.email = true;
-    if (!contactMessage) newErrorState.message = true;
-
-    //If all fields valid: Try to send email and form fields, catch any error
-    if (!newErrorState.name && !newErrorState.email && !newErrorState.message) {
-      try {
-        await sendEmail();
-        setContactName("");
-        setContactEmail("");
-        setContactMessage("");
-        setSuccess(true);
-      } catch (error) {
-        console.log(error)
-        window.alert("Something went wrong, please try again soon" + '\n' + JSON.stringify(error))
-      }
-    }
-    setError(newErrorState);
-  };
-
-  const sendEmail = () => {
-    return new Promise((resolve, reject) => {
-      //These IDs come from the emailjs website (using screenumass gmail account)
-      emailjs.sendForm('service_k7xidgk', 'template_15g5s3y', form.current, 'VU9U1vX9cAro8XtUK')
-        .then((result) => {
-          // console.log(result.text);
-          resolve(result);
-        })
-        .catch((error) => {
-          reject(error);
-        });
-    });
-  };
-
-
 
   return (
     <main>
@@ -246,70 +171,7 @@ export default function About() {
               <Typography display={"inline"} variant="body1">If you would like to send an attachment, feel free to email us directly at&nbsp;</Typography>
               <Link display={"inline"} href="mailto:encode-screen@googlegroups.com" target="_blank" rel="noopener noreferrer">encode&#8209;screen@googlegroups.com</Link>
             </Box>
-            <Box
-              component="form"
-              ref={form}
-              id="contact-us"
-              sx={{
-                '& > :not(style)': { width: '50ch' },
-              }}
-              noValidate
-              autoComplete="off"
-            >
-              <TextField
-                required
-                value={contactName}
-                onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
-                  setContactName(event.target.value);
-                }}
-                error={error.name}
-                name="user_name"
-                type="text"
-                sx={{ display: 'block', mb: 1 }}
-                id="outlined-basic"
-                label="Name"
-                variant="outlined"
-              />
-              <TextField
-                required
-                value={contactEmail}
-                onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
-                  setContactEmail(event.target.value);
-                }}
-                error={error.email || contactEmail !== '' && !isValidEmail(contactEmail)}
-                helperText={error.email && "Please enter a valid email"}
-                name="user_email"
-                type="email"
-                sx={{ display: 'block', mb: 1 }}
-                id="outlined-basic"
-                label="Email"
-                variant="outlined"
-              />
-              <TextField
-                required
-                value={contactMessage}
-                onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
-                  setContactMessage(event.target.value);
-                }}
-                error={error.message}
-                name="message"
-                type="text"
-                fullWidth
-                rows={4}
-                sx={{ display: 'block' }}
-                multiline id="outlined-basic"
-                label="Message"
-                variant="outlined"
-              />
-              <Button
-                sx={{ mt: 1 }}
-                variant="contained"
-                onClick={handleSubmit}
-              >
-                Submit
-              </Button>
-            </Box>
-            {success && <Typography>Submitted successfully, thank you!</Typography>}
+            <ContactForm />
           </Grid>
         </Grid>
       </Grid>
